chore: Move info-related code under info directory
