+++ src/state/info/queries/pools/topPools.ts
@@ -2,7 +2,7 @@ import { useState, useEffect } from 'react'
 import { request, gql } from 'graphql-request'
 import { INFO_CLIENT } from 'config/constants/endpoints'
 import { TOKEN_BLACKLIST } from 'config/constants/info'
-import { getDeltaTimestamps } from 'utils/infoQueryHelpers'
+import { getDeltaTimestamps } from 'views/Info/utils/infoQueryHelpers'
 
 interface TopPoolsResponse {
   pairDayDatas: {
