+++ src/state/info/queries/tokens/priceData.ts
@@ -1,7 +1,7 @@
 import { getUnixTime } from 'date-fns'
 import { gql } from 'graphql-request'
-import { getBlocksFromTimestamps } from 'hooks/useBlocksFromTimestamps'
-import { multiQuery } from 'utils/infoQueryHelpers'
+import { getBlocksFromTimestamps } from 'views/Info/hooks/useBlocksFromTimestamps'
+import { multiQuery } from 'views/Info/utils/infoQueryHelpers'
 import { PriceChartEntry } from 'state/info/types'
 import { INFO_CLIENT } from 'config/constants/endpoints'
