+++ src/views/Info/Overview/index.tsx
@@ -4,11 +4,11 @@ import { Flex, Box, Text, Heading, Card, Skeleton } from '@pancakeswap/uikit'
 import { format, fromUnixTime } from 'date-fns'
 import { useTranslation } from 'contexts/Localization'
 import Page from 'components/Layout/Page'
-import LineChart from 'components/InfoCharts/LineChart'
-import TokenTable from 'components/InfoTables/TokensTable'
-import PoolTable from 'components/InfoTables/PoolsTable'
-import { formatAmount } from 'utils/formatInfoNumbers'
-import BarChart from 'components/InfoCharts/BarChart'
+import LineChart from 'views/Info/components/InfoCharts/LineChart'
+import TokenTable from 'views/Info/components/InfoTables/TokensTable'
+import PoolTable from 'views/Info/components/InfoTables/PoolsTable'
+import { formatAmount } from 'views/Info/utils/formatInfoNumbers'
+import BarChart from 'views/Info/components/InfoCharts/BarChart'
 import {
   useAllPoolData,
   useAllTokenData,
@@ -16,7 +16,7 @@ import {
   useProtocolData,
   useProtocolTransactions,
 } from 'state/info/hooks'
-import TransactionTable from 'components/InfoTables/TransactionsTable'
+import TransactionTable from 'views/Info/components/InfoTables/TransactionsTable'
 
 export const ChartCardsContainer = styled(Flex)`
   justify-content: space-between;
