+++ src/views/Info/Tokens/index.tsx
@@ -1,11 +1,11 @@
 import React, { useMemo, useEffect } from 'react'
 import { Text, Heading, Card } from '@pancakeswap/uikit'
 import Page from 'components/Layout/Page'
-import TokenTable from 'components/InfoTables/TokensTable'
+import TokenTable from 'views/Info/components/InfoTables/TokensTable'
 import { useAllTokenData, useTokenDatas } from 'state/info/hooks'
 import { useWatchlistTokens } from 'state/user/hooks'
 import { useTranslation } from 'contexts/Localization'
-import TopTokenMovers from 'components/TopTokenMovers'
+import TopTokenMovers from 'views/Info/components/TopTokenMovers'
 
 const TokensOverview: React.FC = () => {
   const { t } = useTranslation()
