+++ src/views/Info/components/ChartLoaders/BarChartLoaderSVG.tsx
