+++ src/views/Info/components/ChartLoaders/CandleChartLoaderSVG.tsx
