+++ src/views/Info/components/ChartLoaders/LineChartLoaderSVG.tsx
