+++ src/views/Info/components/ChartLoaders/index.tsx
