+++ src/views/Info/components/CurrencyLogo/index.tsx
