+++ src/views/Info/components/InfoCharts/ChartCard/index.tsx
@@ -1,11 +1,11 @@
 import React, { useMemo, useState } from 'react'
 import { Text, Box, Card, Flex, Skeleton } from '@pancakeswap/uikit'
-import LineChart from 'components/InfoCharts/LineChart'
-import BarChart from 'components/InfoCharts/BarChart'
-import CandleChart from 'components/InfoCharts/CandleChart'
+import LineChart from 'views/Info/components/InfoCharts/LineChart'
+import BarChart from 'views/Info/components/InfoCharts/BarChart'
+import CandleChart from 'views/Info/components/InfoCharts/CandleChart'
 import { TabToggleGroup, TabToggle } from 'components/TabToggle'
 import { useTranslation } from 'contexts/Localization'
-import { formatAmount } from 'utils/formatInfoNumbers'
+import { formatAmount } from 'views/Info/utils/formatInfoNumbers'
 import { ChartEntry, TokenData, PriceChartEntry } from 'state/info/types'
 import { format, fromUnixTime } from 'date-fns'
