+++ src/views/Info/components/InfoTables/TokensTable.tsx
@@ -3,9 +3,9 @@ import styled from 'styled-components'
 import { Text, Flex, Box, Skeleton, useMatchBreakpoints, ArrowBackIcon, ArrowForwardIcon } from '@pancakeswap/uikit'
 import { TokenData } from 'state/info/types'
 import { Link } from 'react-router-dom'
-import { CurrencyLogo } from 'components/CurrencyLogo'
-import { formatAmount } from 'utils/formatInfoNumbers'
-import Percent from 'components/Percent'
+import { CurrencyLogo } from 'views/Info/components/CurrencyLogo'
+import { formatAmount } from 'views/Info/utils/formatInfoNumbers'
+import Percent from 'views/Info/components/Percent'
 import { useTranslation } from 'contexts/Localization'
 import { ClickableColumnHeader, TableWrapper, PageButtons, Arrow, Break } from './shared'
