+++ src/views/Info/components/InfoTables/shared.tsx
