+++ src/views/Info/components/Percent/index.tsx
