+++ src/views/Info/components/SaveIcon/index.tsx
