+++ src/views/Info/components/TopTokenMovers/index.tsx
@@ -4,9 +4,9 @@ import { Text, Flex, Box, Card } from '@pancakeswap/uikit'
 import { Link } from 'react-router-dom'
 import { useAllTokenData } from 'state/info/hooks'
 import { TokenData } from 'state/info/types'
-import { CurrencyLogo } from 'components/CurrencyLogo'
-import { formatAmount } from 'utils/formatInfoNumbers'
-import Percent from 'components/Percent'
+import { CurrencyLogo } from 'views/Info/components/CurrencyLogo'
+import { formatAmount } from 'views/Info/utils/formatInfoNumbers'
+import Percent from 'views/Info/components/Percent'
 import { useTranslation } from 'contexts/Localization'
 
 const CardWrapper = styled(Link)`
