+++ src/views/Info/utils/formatInfoNumbers.ts
