+++ src/views/Info/utils/infoDataHelpers.ts
