+++ src/views/Info/utils/infoQueryHelpers.ts
